+++ lib/searchkick/query.rb
@@ -19,7 +19,7 @@ module Searchkick
         :boost_by, :boost_by_distance, :boost_where, :conversions, :conversions_term, :debug, :emoji, :exclude, :execute, :explain,
         :fields, :highlight, :includes, :index_name, :indices_boost, :limit, :load,
         :match, :misspellings, :model_includes, :offset, :operator, :order, :padding, :page, :per_page, :profile,
-        :request_params, :routing, :select, :similar, :smart_aggs, :suggest, :track, :type, :where]
+        :request_params, :routing, :scope_results, :select, :similar, :smart_aggs, :suggest, :track, :type, :where]
       raise ArgumentError, "unknown keywords: #{unknown_keywords.join(", ")}" if unknown_keywords.any?
 
       term = term.to_s
@@ -98,7 +98,7 @@ module Searchkick
       # no easy way to tell which host the client will use
       host = Searchkick.client.transport.hosts.first
       credentials = host[:user] || host[:password] ? "#{host[:user]}:#{host[:password]}@" : nil
-      "curl #{host[:protocol]}://#{credentials}#{host[:host]}:#{host[:port]}/#{CGI.escape(index)}#{type ? "/#{type.map { |t| CGI.escape(t) }.join(',')}" : ''}/_search?pretty -d '#{query[:body].to_json}'"
+      "curl #{host[:protocol]}://#{credentials}#{host[:host]}:#{host[:port]}/#{CGI.escape(index)}#{type ? "/#{type.map { |t| CGI.escape(t) }.join(',')}" : ''}/_search?pretty -H 'Content-Type: application/json' -d '#{query[:body].to_json}'"
     end
 
     def handle_response(response)
@@ -112,7 +112,9 @@ module Searchkick
         json: !@json.nil?,
         match_suffix: @match_suffix,
         highlighted_fields: @highlighted_fields || [],
-        misspellings: @misspellings
+        misspellings: @misspellings,
+        term: term,
+        scope_results: options[:scope_results]
       }
 
       if options[:debug]
@@ -455,12 +457,16 @@ module Searchkick
           where[:type] = [options[:type] || klass].flatten.map { |v| searchkick_index.klass_document_type(v, true) }
         end
 
-        # filters
+        # start everything as efficient filters
+        # move to post_filters as aggs demand
         filters = where_filters(where)
-        set_filters(payload, filters) if filters.any?
+        post_filters = []
 
         # aggregations
-        set_aggregations(payload) if options[:aggs]
+        set_aggregations(payload, filters, post_filters) if options[:aggs]
+
+        # filters
+        set_filters(payload, filters, post_filters)
 
         # suggestions
         set_suggestions(payload, options[:suggest]) if options[:suggest]
@@ -654,12 +660,11 @@ module Searchkick
       @highlighted_fields = payload[:highlight][:fields].keys
     end
 
-    def set_aggregations(payload)
+    def set_aggregations(payload, filters, post_filters)
       aggs = options[:aggs]
       payload[:aggs] = {}
 
       aggs = Hash[aggs.map { |f| [f, {}] }] if aggs.is_a?(Array) # convert to more advanced syntax
-
       aggs.each do |field, agg_options|
         size = agg_options[:limit] ? agg_options[:limit] : 1_000
         shared_agg_options = agg_options.slice(:order, :min_doc_count)
@@ -704,6 +709,17 @@ module Searchkick
         where = {}
         where = (options[:where] || {}).reject { |k| k == field } unless options[:smart_aggs] == false
         agg_filters = where_filters(where.merge(agg_options[:where] || {}))
+
+        # only do one level comparison for simplicity
+        filters.select! do |filter|
+          if agg_filters.include?(filter)
+            true
+          else
+            post_filters << filter
+            false
+          end
+        end
+
         if agg_filters.any?
           payload[:aggs][field] = {
             filter: {
@@ -719,14 +735,16 @@ module Searchkick
       end
     end
 
-    def set_filters(payload, filters)
-      if options[:aggs]
+    def set_filters(payload, filters, post_filters)
+      if post_filters.any?
         payload[:post_filter] = {
           bool: {
-            filter: filters
+            filter: post_filters
           }
         }
-      else
+      end
+
+      if filters.any?
         # more efficient query if no aggs
         payload[:query] = {
           bool: {
@@ -770,7 +788,7 @@ module Searchkick
           if value.is_a?(Hash)
             value.each do |op, op_value|
               case op
-              when :within, :bottom_right
+              when :within, :bottom_right, :bottom_left
                 # do nothing
               when :near
                 filters << {
@@ -802,6 +820,15 @@ module Searchkick
                     field => {
                       top_left: location_value(op_value),
                       bottom_right: location_value(value[:bottom_right])
+                    }
+                  }
+                }
+              when :top_right
+                filters << {
+                  geo_bounding_box: {
+                    field => {
+                      top_right: location_value(op_value),
+                      bottom_left: location_value(value[:bottom_left])
                     }
                   }
                 }
@@ -887,7 +914,7 @@ module Searchkick
           field_value_factor: {
             field: field,
             factor: value[:factor].to_f,
-            modifier: log ? "ln2p" : nil
+            modifier: value[:modifier] || (log ? "ln2p" : nil)
           }
         }