+++ test/multi_search_test.rb
@@ -33,4 +33,12 @@ class MultiSearchTest < Minitest::Test
     Searchkick.multi_search([products], retry_misspellings: true)
     assert_equal ["abc", "abd"], products.map(&:name)
   end
+
+  def test_error
+    products = Product.search("*", order: {bad_column: :asc}, execute: false)
+    Searchkick.multi_search([products])
+    assert products.error
+    error = assert_raises(Searchkick::Error) { products.results }
+    assert_equal error.message, "Query error - use the error method to view it"
+  end
 end