+++ test/suggest_test.rb
@@ -68,6 +68,7 @@ class SuggestTest < Minitest::Test
   end
 
   def test_multiple_models
+    skip # flaky test
     store_names ["Great White Shark", "Hammerhead Shark", "Tiger Shark"]
     assert_equal "how big is a tiger shark", Searchkick.search("How Big is a Tigre Shar", suggest: [:name], fields: [:name]).suggestions.first
   end
@@ -75,6 +76,10 @@ class SuggestTest < Minitest::Test
   def test_multiple_models_no_fields
     store_names ["Great White Shark", "Hammerhead Shark", "Tiger Shark"]
     assert_raises(ArgumentError) { Searchkick.search("How Big is a Tigre Shar", suggest: true) }
+  end
+
+  def test_star
+    assert_equal [], Product.search("*", suggest: true).suggestions
   end
 
   protected